refactor: move obs widget to client component
